SmartEdu/requirements.txt
# Project Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Flask-Login==0.6.2
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.0
flask-cors==4.0.0
Flask-WTF==1.1.1
Flask-Migrate==4.0.0
alembic==1.11.1
markdown